requirements.txt
paho-mqtt==1.6.1
flask==2.3.2
requests==2.31.0